Move ProgramFromDictionary to 'dictionary' package
